@@ -4,11 +4,14 @@
   "description": "Useful yup schema shortcuts",
   "main": "dist/index.js",
   "scripts": {
-    "build": "webpack --config webpack.config.js",
-    "prepublish": "npm run build",
-    "test": "npm run build && mocha --compilers js:babel-core/register",
-    "gendocs": "jsdoc -c .jsdoc && jsdoc2md -c .jsdoc --files lib/* >> docs/INDEX.md",
-    "gendocs:explain": "jsdoc -X -c .jsdoc"
+    "build": "webpack --config webpack.config.babel.js",
+    "prepare": "npm run build",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "flow": "flow",
+    "lint": "eslint *.js src/lib",
+    "docs": "jsdoc -c jsdoc.json",
+    "docs:explain": "jsdoc -X -c jsdoc.json"
   },
   "repository": {
     "type": "git",
@@ -22,16 +25,26 @@
   "license": "MIT",
   "devDependencies": {
     "babel-core": "^6.25.0",
-    "babel-loader": "^7.0.0",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "^7.1.1",
     "babel-preset-es2015": "^6.24.1",
+    "babel-preset-flow": "^6.23.0",
     "babel-preset-stage-1": "^6.24.1",
-    "chai": "^4.0.2",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.0.2",
+    "eslint-plugin-babel": "^4.1.1",
+    "eslint-plugin-flowtype": "^2.34.1",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jest": "^20.0.3",
+    "eslint-plugin-jsx-a11y": "^5.0.1",
+    "eslint-plugin-react": "^7.0.1",
+    "flow-bin": "^0.49.1",
     "ink-docstrap": "^1.3.0",
+    "jest": "^20.0.4",
     "jsdoc": "^3.4.3",
     "jsdoc-babel": "^0.3.0",
-    "jsdoc-to-markdown": "^3.0.0",
-    "mocha": "^3.4.2",
-    "webpack": "^2.6.1",
+    "precommit-hook-eslint": "^3.0.0",
+    "webpack": "^3.0.0",
     "webpack-node-externals": "^1.6.0"
   },
   "dependencies": {
@@ -39,5 +52,10 @@
     "moment": "^2.18.1",
     "social-insurance-number": "^0.0.2",
     "yup": "^0.21.3"
-  }
+  },
+  "pre-commit": [
+    "lint",
+    "flow",
+    "test"
+  ]
 }
