@@ -1,3 +1,5 @@
+// @flow
+
 /**
  * Useful yup schema shortcuts
  */
@@ -28,7 +30,7 @@ const password = yup
  * Validates a second password.
  * @param passwordField
  */
-const passwordSecond = passwordField => yup.string().when(passwordField, v => {
+const passwordSecond = (passwordField: string) => yup.string().when(passwordField, v => {
 	if (!v) return yup.string().required('Please confirm your password');
 	return yup.string().matches(new RegExp(`^${escapeRegExp(v)}$`), 'Passwords must match');
 });
@@ -75,7 +77,7 @@ class MomentSchema extends mixed {
 		});
 	}
 
-	_typeCheck(v) {
+	_typeCheck(v: Object) {
 		return moment.isMoment(v) && v.isValid();
 	}
 }
@@ -111,5 +113,5 @@ export default {
 
 /**
  * Generic Yup Schema
- * @typedef YupSchema
+ * @typedef {Object} YupSchema
  */
